# Conflicts:
#	make/build_targets.mk

+++ php/drupal/set-permissions.sh
@@ -5,14 +5,19 @@
 function set_permissions {
   FILES_PATH=/app/${WEBROOT:-public}/sites/default/files/
 
-  echo "- Delete css and js folder from public files"
-  sudo rm -rf "${FILES_PATH}css" "${FILES_PATH}js"
+  if [ -d "${FILES_PATH}" ]; then
 
-  echo "- Set ownership of ${FILES_PATH} to www-data"
-  sudo chown -R www-data:www-data "${FILES_PATH}"
+    echo "- Delete css and js folder from public files"
+    sudo rm -rf "${FILES_PATH}css" "${FILES_PATH}js"
 
-  echo "- Set group permissions for ${FILES_PATH}"
-  sudo chmod g+rwx "${FILES_PATH}"
+    echo "- Set ownership of ${FILES_PATH} to www-data"
+    sudo chown -R www-data:www-data "${FILES_PATH}"
+
+    echo "- Set group permissions for ${FILES_PATH}"
+    sudo chmod g+rwx "${FILES_PATH}"
+
+  fi
+
 }
 
 set_permissions &
